ref: 대시보드 서비스 컨트롤러 통합 리팩토링
